requirements.txt
groq==0.11.0
protobuf==5.28.2
python-dotenv==1.0.1
streamlit==1.38.0
